+++ js-miniapp-bridge/test/test.spec.ts
@@ -6,6 +6,15 @@ import {
   CustomPermissionName,
   CustomPermissionStatus,
 } from '../src/types/custom-permissions';
+import {
+  errorTypesDescriptions,
+  MiniAppErrorType,
+  parseMiniAppError,
+  AudienceNotSupportedError,
+  ScopesNotSupportedError,
+  AuthorizationFailureError,
+  MiniAppError,
+} from '../src';
 
 /* tslint:disable:no-any */
 const window: any = {};
@@ -71,6 +80,96 @@ describe('execErrorCallback', () => {
       Bridge.mabMessageQueue.unshift(callback);
       Bridge.MiniAppBridge.prototype.execErrorCallback(callback.id, '');
     });
+  });
+});
+
+describe('getToken', () => {
+  it('will parse the AccessToken JSON response', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+    mockExecutor.exec.callsArgWith(
+      2,
+      '{ "token": "test", "validUntil": 0, "scopes": { "audience": "AUD", "scopes": ["SCO1","SCO2"]} }'
+    );
+
+    return expect(
+      bridge.getAccessToken('AUD', ['SCO1', 'SCO2'])
+    ).to.eventually.deep.equal({
+      token: 'test',
+      validUntil: new Date(0),
+      scopes: {
+        audience: 'AUD',
+        scopes: ['SCO1', 'SCO2'],
+      },
+    });
+  });
+
+  it('will parse the AccessToken AudienceNotSupportedError JSON response', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+    mockExecutor.exec.callsArgWith(
+      3,
+      '{ "type": "AudienceNotSupportedError", "message": null }'
+    );
+
+    return expect(bridge.getAccessToken('AUDIENCE', ['SCOPE1', 'SCOPE2']))
+      .to.eventually.be.rejected.and.be.an.instanceof(AudienceNotSupportedError)
+      .and.have.property(
+        'message',
+        errorTypesDescriptions.get(MiniAppErrorType.AudienceNotSupportedError)
+      );
+  });
+
+  it('will parse the AccessToken ScopesNotSupportedError JSON response', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+    mockExecutor.exec.callsArgWith(3, '{ "type": "ScopesNotSupportedError" }');
+
+    return expect(bridge.getAccessToken('AUDIENCE', ['SCOPE1', 'SCOPE2']))
+      .to.eventually.be.rejected.and.be.an.instanceof(ScopesNotSupportedError)
+      .and.have.property(
+        'message',
+        errorTypesDescriptions.get(MiniAppErrorType.ScopesNotSupportedError)
+      );
+  });
+
+  it('will parse the AccessToken AuthorizationFailureError JSON response', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+    mockExecutor.exec.callsArgWith(
+      3,
+      '{ "type": "AuthorizationFailureError", "message": "test message" }'
+    );
+
+    return expect(bridge.getAccessToken('AUDIENCE', ['SCOPE1', 'SCOPE2']))
+      .to.eventually.be.rejected.and.be.an.instanceof(AuthorizationFailureError)
+      .and.have.property('message', 'test message');
+  });
+
+  it('will parse the AccessToken error JSON response', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+    mockExecutor.exec.callsArgWith(
+      3,
+      '{ "type": "test", "message": "test message" }'
+    );
+
+    return expect(bridge.getAccessToken('AUDIENCE', ['SCOPE1', 'SCOPE2']))
+      .to.eventually.be.rejected.and.be.an.instanceof(MiniAppError)
+      .and.to.include({ name: 'test', message: 'test message' });
+  });
+
+  it('will parse the AccessToken error JSON with no type response', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+    mockExecutor.exec.callsArgWith(3, '{ "message": "test message" }');
+
+    return expect(bridge.getAccessToken('AUDIENCE', ['SCOPE1', 'SCOPE2']))
+      .to.eventually.be.rejected.and.be.an.instanceof(MiniAppError)
+      .and.to.include({ message: 'test message' });
+  });
+
+  it('will still send an error if JSON error parsing fails', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+    mockExecutor.exec.callsArgWith(3, 'an error occurred');
+
+    return expect(
+      bridge.getAccessToken('AUDIENCE', ['SCOPE1', 'SCOPE2'])
+    ).to.eventually.be.rejected.and.to.equal('an error occurred');
   });
 });