+++ js-miniapp-sdk/test/miniapp.spec.ts
@@ -9,6 +9,13 @@ import {
   CustomPermissionStatus,
   ScreenOrientation,
   MessageToContact,
+  AudienceNotSupportedError,
+  ScopesNotSupportedError,
+  AuthorizationFailureError,
+  MiniAppError,
+  parseMiniAppError,
+  errorTypesDescriptions,
+  MiniAppErrorType,
 } from '../../js-miniapp-bridge/src';
 import { MiniApp } from '../src/miniapp';
 
@@ -330,6 +337,98 @@ describe('getAccessToken', () => {
     return expect(
       miniApp.user.getAccessToken('AUDIENCE', ['SCOPE1', 'SCOPE2'])
     ).to.eventually.equal(response);
+  });
+
+  it('should retrieve AudienceNotSupportedError response from the MiniAppBridge once there is an audience error', () => {
+    const json = parseMiniAppError('{"type": "AudienceNotSupportedError"}');
+    const error = new AudienceNotSupportedError(json);
+
+    expect(error.message).to.equal(
+      errorTypesDescriptions.get(MiniAppErrorType.AudienceNotSupportedError)
+    );
+
+    expect(error.name).to.equal(MiniAppErrorType.AudienceNotSupportedError);
+
+    window.MiniAppBridge.getAccessToken.resolves(error);
+    return expect(
+      miniApp.user.getAccessToken('AUDIENCE', ['SCOPE1', 'SCOPE2'])
+    ).to.eventually.equal(error);
+  });
+
+  it('should retrieve ScopesNotSupportedError response from the MiniAppBridge once there is an scope error', () => {
+    const json = parseMiniAppError('{"type": "ScopesNotSupportedError"}');
+    const error = new ScopesNotSupportedError(json);
+
+    expect(error.message).to.equal(
+      errorTypesDescriptions.get(MiniAppErrorType.ScopesNotSupportedError)
+    );
+
+    expect(error.name).to.equal(MiniAppErrorType.ScopesNotSupportedError);
+
+    window.MiniAppBridge.getAccessToken.resolves(error);
+    return expect(
+      miniApp.user.getAccessToken('AUDIENCE', ['SCOPE1', 'SCOPE2'])
+    ).to.eventually.equal(error);
+  });
+
+  it('should retrieve AuthorizationFailureError response from the MiniAppBridge once there is an authorization error', () => {
+    const json = parseMiniAppError(
+      '{"type": "AuthorizationFailureError", "message": "authorization failed"}'
+    );
+    const error = new AuthorizationFailureError(json);
+
+    expect(error.message).to.equal('authorization failed');
+
+    expect(error.name).to.equal(MiniAppErrorType.AuthorizationFailureError);
+
+    window.MiniAppBridge.getAccessToken.resolves(error);
+    return expect(
+      miniApp.user.getAccessToken('AUDIENCE', ['SCOPE1', 'SCOPE2'])
+    ).to.eventually.equal(error);
+  });
+
+  it('should retrieve MiniAppError response from the MiniAppBridge once there is an error with no type', () => {
+    const json = parseMiniAppError('{"message": "authorization failed"}');
+    const error = new MiniAppError(json);
+
+    expect(error.message).to.equal('authorization failed');
+
+    expect(error.name).to.equal(undefined);
+
+    window.MiniAppBridge.getAccessToken.resolves(error);
+    return expect(
+      miniApp.user.getAccessToken('AUDIENCE', ['SCOPE1', 'SCOPE2'])
+    ).to.eventually.equal(error);
+  });
+
+  it('should retrieve MiniAppError response from the MiniAppBridge once there is an error with an unknown type', () => {
+    const json = parseMiniAppError(
+      '{"type": "Other", "message": "authorization failed"}'
+    );
+    const error = new MiniAppError(json);
+
+    expect(error.message).to.equal('authorization failed');
+
+    expect(error.name).to.equal('Other');
+
+    window.MiniAppBridge.getAccessToken.resolves(error);
+    return expect(
+      miniApp.user.getAccessToken('AUDIENCE', ['SCOPE1', 'SCOPE2'])
+    ).to.eventually.equal(error);
+  });
+
+  it('should retrieve MiniAppError response from the MiniAppBridge once there is an error with no type and no message', () => {
+    const json = parseMiniAppError('{}');
+    const error = new MiniAppError(json);
+
+    expect(error.message).to.equal(undefined);
+
+    expect(error.name).to.equal(undefined);
+
+    window.MiniAppBridge.getAccessToken.resolves(error);
+    return expect(
+      miniApp.user.getAccessToken('AUDIENCE', ['SCOPE1', 'SCOPE2'])
+    ).to.eventually.equal(error);
   });
 });