+++ js-miniapp-sdk/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## CHANGELOG
 
+### 1.10.0 (WIP)
+- **Feature:** Added `MiniAppError` class for a better bridge errors management [See here](README.md#Errors-management).
+
 ### 1.9.0 (2021-05-27)
 - **Feature:** Added `Send Messsage` custom permission for sending message to specific contact. [See here](README.md#Request-Permissions).
+- **Feature:** Added `MiniAppError` class for a better bridge errors management [See here](README.md#Errors-management).
 
 ### 1.8.0 (2021-04-22)